CollegeCampusApp/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==3.0.10
requests==2.31.0
python-dotenv==1.0.0
bcrypt==5.0.0
